+++ content/switches.py
@@ -9,7 +9,7 @@ from pyswitch.hardware.Hardware import Hardware
 # from pyswitch.misc import Colors
 
 from pyswitch.clients.kemper import KemperEffectSlot
-from pyswitch.controller.PedalController import PedalAction
+from pyswitch.controller.InputAction import InputAction
 from display import DISPLAY_FOOTER_1, DISPLAY_FOOTER_2, DISPLAY_HEADER_1, DISPLAY_HEADER_2
 
 from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
@@ -212,12 +212,12 @@ Switches = [
 
 #################################################################################
 
-# Expression Pedals (None if not used)
-Pedals = [
+# Expression Pedals and other inputs (None if not used)
+Inputs = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_EXP_PEDAL_1,
         "actions": [
-            PedalAction(
+            InputAction(
                 mapping = MAPPING_RIG_VOLUME()
             )
         ]