+++ examples/Kemper Profiler Player/MIDICaptain 10/Arek77/switches.py
@@ -204,4 +204,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None