+++ examples/Kemper Profiler Player/MIDICaptain 10/Erikcb/switches.py
@@ -129,4 +129,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None