+++ examples/Kemper Profiler Player/MIDICaptain Mini 6/CurryFlurry/switches.py
@@ -88,4 +88,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None