+++ examples/Kemper Profiler Player/MIDICaptain Mini 6/gstrotmann/switches.py
@@ -84,4 +84,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None