+++ examples/Kemper Profiler Player/MIDICaptain Mini 6/kitarist/switches.py
@@ -114,4 +114,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None