+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Bank Up Down/switches.py
@@ -61,4 +61,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None