+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Custom Bank Names/switches.py
@@ -90,4 +90,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None