+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/FX States and Amp Name/switches.py
@@ -63,4 +63,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None