+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/FX States and Bank Change/switches.py
@@ -71,4 +71,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None