+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Freeze and Tap Tempo/switches.py
@@ -76,4 +76,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None