+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Rig and Bank Select plus Morph State/switches.py
@@ -81,4 +81,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None