+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Rig and Bank Select/switches.py
@@ -78,4 +78,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None