+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Tuner and RigVolumeBoost/switches.py
@@ -65,4 +65,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None