+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Tuner and Tap Tempo and Morph Button/switches.py
@@ -68,4 +68,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None