+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/tunetown KBDN/switches.py
@@ -75,4 +75,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None