+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/tunetown Session with Looper and Pages/switches.py
@@ -165,4 +165,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None