+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/tunetown Session/switches.py
@@ -87,4 +87,5 @@ Switches = [
 
 ######################################################################
 
-Pedals = None
+# Expression pedals and other inputs
+Inputs = None